requirements.txt
firecloud>=0.16.9
firecloud-dalmatian>=0.0.6
google-cloud-storage>=1.13.2
PyYAML>=5.1
agutil>=4.0.2
Flask-Cors>=3.0.7
crayons>=0.2.0
connexion>=2.2.0
oauth2client>=4.1.3
requests>=2.21.0
googleapis-common-protos>=1.5.3
google-auth>=1.4.2
google-cloud-kms>=1.0.0
cryptography>=2.3.0
pyOpenSSL>=19.0.0
protobuf>=3.6.0
